+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Mapbox welcomes participation and contributions from everyone.
 
 ### main
+- Add `StepIntersection#duration` (previously available through `StepIntersection#getUnrecognizedProperty("duration")`)
 
 ### v7.7.0 - August 13, 2025
 - Add `RouteLeg#notifications`